Added remarks field, more lots
Remarks field is optional and simply for extra information about the lot. It is intended ot add information, especially if the primary color or part no is the majority, so it is easier to calculate quantity based on that, and simply add  a comment about an extra color or part.